evaluate cycles for UpdatePkey and HeapSwitch
before this commit I tend to wrap the function to be tested, in order to
easily switch on/off profiling.
However, I happened to find that introducing an extra layer of function
call will significantly increase the overhead, as well as making it
very unstable. UpdatePkey has an overhead of ~80 vs ~450 cycles due to
this issue. Anyway, not introducing an extra layer is perfectly right
and correspond to existing literature, too.