archive failed attempt to load PIE
PIEs tend to use static TLS, and rtld collects the modules defining them
and allocate static TLS for them. While static TLS is possible for some
of the libraries, one might have a quite large TLS segment and thus hav-
ing no hope for it. Though still not clear about how static TLS is allo-
cated for DSO, it is hard and complicated, so not now